implemented feature to calculate tax amt, amount and total. Added functionality to remove item button

+++ ShoppingBird.Mobile/ShoppingBird.Mobile/Models/TaxModel.cs
@@ -8,6 +8,6 @@ namespace ShoppingBird.Mobile.Models
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public decimal Percent { get; set; }
+        public double Percent { get; set; }
     }
 }
